django/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Forest Anderson <forestkzanderson@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.3"
"discord.py" = "^1.7.2"
psutil = "^5.8.0"
tabulate = "^0.8.9"
djangorestframework = "^3.12.4"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
black = "^21.5b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
